remove plural modules
