--- lib/outboxer/services/publishers_service.rb
@@ -1,32 +1,0 @@
-module Outboxer
-  module PublishersService
-    module_function
-
-    def all
-      ActiveRecord::Base.connection_pool.with_connection do
-        ActiveRecord::Base.transaction do
-          publishers = Publisher.includes(:signals).all
-
-          publishers.map do |publisher|
-            {
-              id: publisher.id,
-              name: publisher.name,
-              status: publisher.status,
-              settings: publisher.settings,
-              metrics: publisher.metrics,
-              created_at: publisher.created_at.utc,
-              updated_at: publisher.updated_at.utc,
-              signals: publisher.signals.map do |signal|
-                {
-                  id: signal.id,
-                  name: signal.name,
-                  created_at: signal.created_at.utc
-                }
-              end
-            }
-          end
-        end
-      end
-    end
-  end
-end
