+++ spec/lib/outboxer/services/setting_service/create_spec.rb
@@ -1,10 +1,10 @@
 require "rails_helper"
 
 module Outboxer
-  RSpec.describe SettingsService, type: :module do
+  RSpec.describe SettingService, type: :module do
     describe ".create" do
       it "creates setting" do
-        SettingsService.create
+        SettingService.create
 
         expect(Setting.find_by(name: "messages.published.count.historic").value).to eq("0")
         expect(Setting.find_by(name: "messages.failed.count.historic").value).to eq("0")
